Updated the command chain of representable object to be more general
Point3d is the most basic thing, it just has an x, a y, and a z and a few static objects like finding the vector between two points.
A polygon is a collection of points that goes in order around the perimeter of the polygon. It also has a color.
A Viewable object has an array of polygons that can be displayed. It's an interface, so it's just used generally in the viewing environment.
The viewer is the person looking around. They have an x, a y, and a z as well as a theta and a phi which represents the angle that they're looking at the viewing environment.
I changed the Background object to the ViewingEnvironment, as that's a more appropriate name. It will still be the driver class, where everything comes together.