Built the basis for the display of the objects
This was a doozy. Lots of stuff. Basically, I added a method that casts a 3d polygon to 2d space based on the position of a viewer. I also sorted all of the polygons in the viewing environment based on their distance from a viewer. Now, All we need to do is actually draw it, but I've done a lot today so we're gonna hold off.